Move update visit to new controller
